examples/agent/a2ui_agent/samples/general_agent/pyproject.toml
[project]
name = "general-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "agentscope[a2a]",
    "agentscope-runtime",
    "uvicorn",
    "starlette",
    "fastapi",
    "shortuuid",
    "a2ui-agent"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
a2ui-agent = { path = "../../../../../../A2UI/a2a_agents/python/a2ui_agent" }
agentscope = { path = "../../../../../" }